refactor: better split of code
