+++ python_bindings/app.py
@@ -76,7 +76,7 @@ for id in ids:
     deleted_ids.append(id)
     segment.delete(id)
 
-print(f' Third search (No filter): Expect 10 results but different from the first since they were removing')
+print(f' Third search (No filter): Expect 10 results but different from the first since they were removed')
 new_ids, new_scores = segment.search(query, None, 10)
 assert set(new_ids) != ids
 
@@ -86,7 +86,9 @@ for id in new_ids:
     print(f' extracted_doc {extracted_doc}')
     assert extracted_doc.text == f'I am document {id}'
 
-new_docs = DocumentArray([Document(id=str(i), embedding=get_random_numpy(), text=f'I am document {i}', granularity=5, weight=5) for i in range(2000, 2500)])
+
+print(f' No we will add documents using the set_full_payload_document interface to serialize the document that will be loaded by the `DocumentProto` in rust')
+new_docs = DocumentArray([Document(id=str(i), embedding=get_random_numpy(), text=f'I am document {i}', granularity=5, weight=5) for i in range(2000, 2010)])
 
 for doc in new_docs:
     doc.tags['hello'] = 'world'
@@ -105,8 +107,8 @@ field1['match'] = {'keyword': 'world'}
 filter['should'] = [field1]
 
 filtered_ids, filtered_scores = segment.search(query, json.dumps(filter), 1000)
-assert len(filtered_ids) == 500
-assert len(filtered_scores) == 500
+assert len(filtered_ids) == 10
+assert len(filtered_scores) == 10
 
 filter = {}
 field1 = {}
@@ -114,11 +116,11 @@ field1['key'] = 'hello'
 field1['match'] = {'keyword': 'world'}
 field2 = {}
 field2['key'] = 'inner_float'
-field2['match'] = {'integer': 2100}
+field2['match'] = {'integer': 2005}
 filter['should'] = [field1, field2]
 print(f' filtered_ids {len(filtered_ids)}')
 filtered_ids, filtered_scores = segment.search(query, json.dumps(filter), 1000)
 print(f' len {len(filtered_ids)}')
 assert len(filtered_ids) == 1
 assert len(filtered_scores) == 1
-assert filtered_ids[0] == 2100
+assert filtered_ids[0] == 2005
